pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ["ka2jun8 <ka210n8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^0.22.2"
mecab-python3 = "^0.996.5"
gensim = "^3.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
start = "sample:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

